% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epdiff.R
\name{epdiff}
\alias{epdiff}
\title{Explain difference between predictions from a model}
\usage{
epdiff(mod, x1, x2, d)
}
\arguments{
\item{mod}{Model}

\item{x1}{Input 1. Single row data frame or matrix.}

\item{x2}{Input 2. Single row data frame or matrix.}

\item{d}{Max number of input dimensions to use.}
}
\value{
ggplot object showing effect
}
\description{
Explain difference between predictions from a model
}
\examples{
# gbm
m1 <- gbm::gbm(Petal.Width ~ Sepal.Length + Sepal.Width, data=iris, distribution='gaussian')
x1 <- iris[1,]
x2 <- iris[2,]
epdiff(m1, x1, x2)
epdiff(m1, iris[12,], iris[20,])
epdiff(m1, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),])

# Linear model
mod_lm <- lm(Petal.Width ~ Sepal.Length + Sepal.Width, data=iris)
epdiff(mod_lm, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),])
}
