% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epdiff1.R
\name{epdiff1}
\alias{epdiff1}
\title{Explain difference between predictions from a model}
\usage{
epdiff1(mod, x1, x2, d, predictfunc, predictfuncwithmod)
}
\arguments{
\item{mod}{Model}

\item{x1}{Input 1. Single row data frame or matrix.}

\item{x2}{Input 2. Single row data frame or matrix.}

\item{d}{Max number of input dimensions to use.}

\item{predictfunc}{Func to use to predict. Defaults to `predict`.
Takes in model and data.}

\item{predictfuncwithmod}{Func to use to predict. Only takes in data.}
}
\value{
ggplot object showing effect
}
\description{
Explain difference between predictions from a model
}
\examples{
# gbm
m1 <- gbm::gbm(Petal.Width ~ Sepal.Length + Sepal.Width, data=iris, distribution='gaussian')
x1 <- iris[1,]
x2 <- iris[2,]
epdiff1(m1, x1, x2)
epdiff1(m1, iris[12,], iris[20,])
epdiff1(m1, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),])
epdiff1(m1, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),], predictfunc=predict)
epdiff1(m1, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),],
    predictfuncwithmod=function(xx){suppressMessages(predict(m1, xx))})
epdiff1(NA, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),],
    predictfuncwithmod=function(xx){suppressMessages(predict(m1, xx))})

# Linear model
mod_lm <- lm(Petal.Width ~ Sepal.Length + Sepal.Width, data=iris)
epdiff1(mod_lm, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),])
epdiff1(mod_lm, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),], d=3)
mod_lm2 <- lm(Petal.Width ~ Petal.Length + Sepal.Width, data=iris)
epdiff1(mod_lm2, iris[sample(1:nrow(iris), 1),], iris[sample(1:nrow(iris), 1),])
}
